+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.34.3](https://github.com/jmerle/competitive-companion/releases/tag/2.34.3) (2022-06-28)
 - Fixed the parsing of HackerEarth Codemonk problems
 - Fixed the parsing of CodeDrills problems when the browser is narrow
 - Fixed the parsing of Toph sample cases